fix(files): Rename files to camelcase
