+++ src/types/Note.tsx
@@ -1,4 +1,4 @@
-export interface CardModel {
+export interface Note {
   id: number;
   title: string;
   text: string;
